+++ pkg/platform/openstack/kuryr_bootstrap.go
@@ -6,7 +6,6 @@ import (
 	"crypto/x509"
 	b64 "encoding/base64"
 	"fmt"
-	"github.com/openshift/cluster-network-operator/pkg/bootstrap"
 	"k8s.io/api/core/v1"
 	"log"
 	"net"
@@ -20,21 +19,9 @@ import (
 
 	"github.com/gophercloud/gophercloud"
 	"github.com/gophercloud/gophercloud/openstack"
-	"github.com/gophercloud/gophercloud/openstack/identity/v3/tokens"
-	"github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/listeners"
-	"github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/loadbalancers"
-	"github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/monitors"
-	"github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/pools"
-	"github.com/gophercloud/gophercloud/openstack/loadbalancer/v2/providers"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/attributestags"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/layer3/routers"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/security/groups"
 	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/security/rules"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/extensions/subnetpools"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/networks"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/ports"
-	"github.com/gophercloud/gophercloud/openstack/networking/v2/subnets"
 	"github.com/gophercloud/utils/openstack/clientconfig"
+	"github.com/openshift/cluster-network-operator/pkg/bootstrap"
 	"github.com/openshift/cluster-network-operator/pkg/names"
 	"github.com/openshift/cluster-network-operator/pkg/platform/openstack/util/cert"
 
@@ -105,666 +92,6 @@ func GetCloudFromSecret(kubeClient client.Client) (clientconfig.Cloud, error) {
 	}
 
 	return clouds.Clouds[CloudName], nil
-}
-
-// Use Neutron tags feature to tag Neutron resources. resource argument must
-// say what type of resource it is.
-func tagResource(client *gophercloud.ServiceClient, resource, id, tag string) ([]string, error) {
-	tagOpts := attributestags.ReplaceAllOpts{Tags: []string{tag}}
-	return attributestags.ReplaceAll(client, resource, id, tagOpts).Extract()
-}
-
-// Looks for a Neutron network by name and tag, if it does not exist creates it.
-// Will fail if two networks match with name and tag.
-func ensureOpenStackNetwork(client *gophercloud.ServiceClient, name, tag string) (string, error) {
-	page, err := networks.List(client, networks.ListOpts{Name: name, Tags: tag}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get network list")
-	}
-	nets, err := networks.ExtractNetworks(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract networks list")
-	}
-	if len(nets) > 1 {
-		return "", errors.Errorf("found multiple networks matching name %s and tag %s, cannot proceed", name, tag)
-	} else if len(nets) == 1 {
-		return nets[0].ID, nil
-	} else {
-		opts := networks.CreateOpts{
-			Name: name,
-		}
-		netObj, err := networks.Create(client, opts).Extract()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to create network")
-		}
-
-		_, err = tagResource(client, "networks", netObj.ID, tag)
-		if err != nil {
-			return "", errors.Wrap(err, "failed to tag created network")
-		}
-
-		return netObj.ID, nil
-	}
-}
-
-// Looks for a Neutron subnetpool by name and tag, if it does not exist creates it.
-// Will fail if two subnetpools match with name and tag.
-func ensureOpenStackSubnetpool(client *gophercloud.ServiceClient, name, tag string, cidrs []string, prefixLen uint32) (string, error) {
-	page, err := subnetpools.List(client, subnetpools.ListOpts{Name: name, Tags: tag}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get subnetpool list")
-	}
-	sp, err := subnetpools.ExtractSubnetPools(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract subnetpools list")
-	}
-	if len(sp) > 1 {
-		return "", errors.Errorf("found multiple subnetpools matching name %s and tag %s, cannot proceed", name, tag)
-	} else if len(sp) == 1 {
-		// TODO(dulek): Check if it has correct CIDRs.
-		return sp[0].ID, nil
-	} else {
-		opts := subnetpools.CreateOpts{
-			Name:             name,
-			Prefixes:         cidrs,
-			DefaultPrefixLen: int(prefixLen),
-		}
-		subnetpoolObj, err := subnetpools.Create(client, opts).Extract()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to create subnetpool")
-		}
-
-		_, err = tagResource(client, "subnetpools", subnetpoolObj.ID, tag)
-		if err != nil {
-			return "", errors.Wrap(err, "failed to tag created subnetpool")
-		}
-
-		return subnetpoolObj.ID, nil
-	}
-}
-
-// Looks for a Neutron subnet by name and tag. Fails if subnet is not found
-// or multiple subnets match.
-func findOpenStackSubnet(client *gophercloud.ServiceClient, name, tag string) (subnets.Subnet, error) {
-	empty := subnets.Subnet{}
-	page, err := subnets.List(client, subnets.ListOpts{Name: name, Tags: tag}).AllPages()
-	if err != nil {
-		return empty, errors.Wrap(err, "failed to get subnet list")
-	}
-	subnetList, err := subnets.ExtractSubnets(page)
-	if err != nil {
-		return empty, errors.Wrap(err, "failed to extract subnets list")
-	}
-	if len(subnetList) == 1 {
-		return subnetList[0], nil
-	} else if len(subnetList) == 0 {
-		return empty, errors.New("subnet not found")
-	} else {
-		return empty, errors.New("multiple matching subnets")
-	}
-}
-
-// Looks for a Neutron subnet by name, tag, network ID, CIDR and gateway IP,
-// if it does not exist creates it using allocationRanges as allocation pools.
-// Will fail if two subnets match all the criteria.
-func ensureOpenStackSubnet(client *gophercloud.ServiceClient, name, tag, netId, cidr, gatewayIp string, allocationRanges []iputil.IPRange) (string, error) {
-	dhcp := false
-	page, err := subnets.List(client, subnets.ListOpts{
-		Name:       name,
-		Tags:       tag,
-		NetworkID:  netId,
-		CIDR:       cidr,
-		GatewayIP:  gatewayIp,
-		IPVersion:  4,
-		EnableDHCP: &dhcp,
-	}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get subnet list")
-	}
-	subnetList, err := subnets.ExtractSubnets(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract subnets list")
-	}
-	if len(subnetList) > 1 {
-		return "", errors.Errorf("found multiple subnets matching name %s and tag %s, cannot proceed", name, tag)
-	} else if len(subnetList) == 1 {
-		return subnetList[0].ID, nil
-	} else {
-		var allocationPools []subnets.AllocationPool
-		for _, r := range allocationRanges {
-			allocationPools = append(allocationPools, subnets.AllocationPool{
-				Start: r.Start.String(), End: r.End.String()})
-		}
-		opts := subnets.CreateOpts{
-			Name:            name,
-			NetworkID:       netId,
-			CIDR:            cidr,
-			GatewayIP:       &gatewayIp,
-			IPVersion:       gophercloud.IPv4,
-			EnableDHCP:      &dhcp,
-			AllocationPools: allocationPools,
-		}
-		subnetObj, err := subnets.Create(client, opts).Extract()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to create subnet")
-		}
-
-		_, err = tagResource(client, "subnets", subnetObj.ID, tag)
-		if err != nil {
-			return "", errors.Wrap(err, "failed to tag created subnet")
-		}
-
-		return subnetObj.ID, nil
-	}
-}
-
-// Looks for a Neutron router by name and tag. Fails if router is not found
-// or multiple routers match.
-func findOpenStackRouter(client *gophercloud.ServiceClient, name, tag string) (routers.Router, error) {
-	empty := routers.Router{}
-	page, err := routers.List(client, routers.ListOpts{Name: name, Tags: tag}).AllPages()
-	if err != nil {
-		return empty, errors.Wrap(err, "failed to get router list")
-	}
-	routerList, err := routers.ExtractRouters(page)
-	if err != nil {
-		return empty, errors.Wrap(err, "failed to extract routers list")
-	}
-
-	if len(routerList) == 1 {
-		return routerList[0], nil
-	} else if len(routerList) == 0 {
-		return empty, errors.New("router not found")
-	} else {
-		return empty, errors.New("multiple matching routers")
-	}
-}
-
-// Returns list of all Neutron ports that belong to a given Neutron router.
-func getOpenStackRouterPorts(client *gophercloud.ServiceClient, routerId string) ([]ports.Port, error) {
-	page, err := ports.List(client, ports.ListOpts{DeviceID: routerId}).AllPages()
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to get port list")
-	}
-	ps, err := ports.ExtractPorts(page)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to extract port list")
-	}
-	return ps, nil
-}
-
-// Returns whether slice ps has a Port that is from subnet subnetId.
-func lookupOpenStackPort(ps []ports.Port, subnetId string) bool {
-	for _, port := range ps {
-		for _, ip := range port.FixedIPs {
-			if ip.SubnetID == subnetId {
-				return true
-			}
-		}
-	}
-	return false
-}
-
-// Adds a subnetId subnet or portId port to routerId router. Will fail if such
-// a connection already exists.
-func ensureOpenStackRouterInterface(client *gophercloud.ServiceClient, routerId string, subnetId, portId *string) error {
-	opts := routers.AddInterfaceOpts{}
-	if subnetId != nil {
-		opts.SubnetID = *subnetId
-	}
-	if portId != nil {
-		opts.PortID = *portId
-	}
-	_, err := routers.AddInterface(client, routerId, opts).Extract()
-	if err != nil {
-		return errors.Wrap(err, "failed to add interface")
-	}
-	return nil
-}
-
-// Looks up OpenStack ports by tag and regexp pattern matched against name.
-// Returns a slice with matched Ports.
-func listOpenStackPortsMatchingPattern(client *gophercloud.ServiceClient, tag string, pattern *regexp.Regexp) ([]ports.Port, error) {
-	page, err := ports.List(client, ports.ListOpts{Tags: tag}).AllPages()
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to get port list")
-	}
-	portList, err := ports.ExtractPorts(page)
-	if err != nil {
-		return nil, errors.Wrap(err, "failed to extract ports list")
-	}
-	result := []ports.Port{}
-	for _, port := range portList {
-		if pattern.MatchString(port.Name) {
-			result = append(result, port)
-		}
-	}
-
-	return result, nil
-}
-
-// Looks for a Neutron Port by name, tag and network ID. If it does not exist
-// creates it. Will fail if multiple ports are matching all criteria.
-func ensureOpenStackPort(client *gophercloud.ServiceClient, name, tag, netId, subnetId, ip string) (string, error) {
-	page, err := ports.List(client, ports.ListOpts{Name: name, Tags: tag, NetworkID: netId}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get port list")
-	}
-	portList, err := ports.ExtractPorts(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract ports list")
-	}
-	if len(portList) > 1 {
-		return "", errors.Errorf("found multiple ports matching name %s, tag %s, cannot proceed", name, tag)
-	} else if len(portList) == 1 {
-		return portList[0].ID, nil
-	} else {
-		opts := ports.CreateOpts{
-			Name:      name,
-			NetworkID: netId,
-			FixedIPs:  []ports.IP{{SubnetID: subnetId, IPAddress: ip}},
-		}
-		portObj, err := ports.Create(client, opts).Extract()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to create port")
-		}
-
-		_, err = tagResource(client, "ports", portObj.ID, tag)
-		if err != nil {
-			return "", errors.Wrap(err, "failed to tag created port")
-		}
-
-		return portObj.ID, nil
-	}
-}
-
-// Looks for a OpenStack security group by name and tag. Fails if SG is not found
-// or multiple SG's match.
-func findOpenStackSgId(client *gophercloud.ServiceClient, name, tag string) (string, error) {
-	page, err := groups.List(client, groups.ListOpts{Name: name, Tags: tag}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get SG list")
-	}
-	sgs, err := groups.ExtractGroups(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract SG list")
-	}
-
-	if len(sgs) == 1 {
-		return sgs[0].ID, nil
-	} else if len(sgs) == 0 {
-		return "", errors.New("SG not found")
-	} else {
-		return "", errors.New("multiple matching SGs")
-	}
-}
-
-// Looks for a OpenStack security group by name and tag. If it does not exist
-// creates it. Will fail if multiple SG's are matching all criteria.
-func ensureOpenStackSg(client *gophercloud.ServiceClient, name, tag string) (string, error) {
-	page, err := groups.List(client, groups.ListOpts{Name: name, Tags: tag}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get SG list")
-	}
-	sgs, err := groups.ExtractGroups(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract SG list")
-	}
-	if len(sgs) > 1 {
-		return "", errors.Errorf("found multiple SG matching name %s, tag %s, cannot proceed", name, tag)
-	} else if len(sgs) == 1 {
-		return sgs[0].ID, nil
-	} else {
-		opts := groups.CreateOpts{
-			Name: name,
-		}
-		sg, err := groups.Create(client, opts).Extract()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to create SG")
-		}
-
-		_, err = tagResource(client, "security-groups", sg.ID, tag)
-		if err != nil {
-			return "", errors.Wrap(err, "failed to tag created SG")
-		}
-
-		return sg.ID, nil
-	}
-}
-
-// Tries to create an OpenStack security group rule on sgId SG. Ignores an
-// error if such rule already exists.
-func ensureOpenStackSgRule(client *gophercloud.ServiceClient, sgId, remotePrefix string, portMin, portMax int, protocol rules.RuleProtocol) error {
-	opts := rules.CreateOpts{
-		SecGroupID:     sgId,
-		EtherType:      rules.EtherType4,
-		Direction:      rules.DirIngress,
-		RemoteIPPrefix: remotePrefix,
-	}
-	// Let's just assume that we're getting passed -1 when we aren't supposed to set those
-	if portMin >= 0 && portMax >= 0 {
-		opts.PortRangeMin = portMin
-		opts.PortRangeMax = portMax
-		opts.Protocol = protocol
-	}
-	_, err := rules.Create(client, opts).Extract()
-	if err != nil {
-		if _, ok := err.(gophercloud.ErrDefault409); ok {
-			// Ignoring 409 Conflict as that means the rule is already there.
-			return nil
-		}
-		return errors.Wrap(err, "failed to create SG rule")
-	}
-	return nil
-}
-
-// Delete a LoadBalancer and all associated resources
-func deleteOpenStackLb(client *gophercloud.ServiceClient, lbId string) error {
-	log.Printf("Deleting Openstack LoadBalancer: %s", lbId)
-	deleteOpts := loadbalancers.DeleteOpts{
-		Cascade: true,
-	}
-	err := loadbalancers.Delete(client, lbId, deleteOpts).ExtractErr()
-	if err != nil {
-		return err
-	}
-
-	err = gophercloud.WaitFor(300, func() (bool, error) {
-		lb, err := loadbalancers.Get(client, lbId).Extract()
-		if err != nil {
-			if _, ok := err.(gophercloud.ErrDefault404); ok {
-				return true, nil
-			}
-			return false, err
-		}
-		if lb.ProvisioningStatus == "DELETED" {
-			return true, nil
-		}
-		return false, nil
-	})
-
-	return err
-}
-
-// Waits up to 5 minutes for OpenStack LoadBalancer to move into ACTIVE
-// provisioning_status. Fails if time runs out.
-func waitForOpenStackLb(client *gophercloud.ServiceClient, lbId string) error {
-	err := gophercloud.WaitFor(300, func() (bool, error) {
-		lb, err := loadbalancers.Get(client, lbId).Extract()
-		if err != nil {
-			return false, err
-		}
-
-		return lb.ProvisioningStatus == "ACTIVE", nil
-	})
-
-	return err
-}
-
-// Looks for a Octavia load balancer by name, address and subnet ID. If it does
-// not exist creates it. Will fail if multiple LB's are matching all criteria.
-func ensureOpenStackLb(client *gophercloud.ServiceClient, name, vipAddress, vipSubnetId, tag string) (string, error) {
-	// We need to figure out if Octavia supports tags and use description field if it's too old. To do that
-	// we list available API versions and look for 2.5. This is because we support Queens and Rocky releases of
-	// OpenStack and those were before tags got implemented.
-	// TODO(dulek): This workaround can be removed once we stop supporting Queens and Rocky OpenStack releases.
-	octaviaTagSupport, err := IsOctaviaVersionSupported(client, MinOctaviaVersionWithTagSupport)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to determine if Octavia supports tags")
-	}
-
-	opts := loadbalancers.ListOpts{
-		Name:        name,
-		VipAddress:  vipAddress,
-		VipSubnetID: vipSubnetId,
-	}
-	if octaviaTagSupport {
-		opts.Tags = []string{tag}
-	} else {
-		opts.Description = tag
-	}
-
-	page, err := loadbalancers.List(client, opts).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get LB list")
-	}
-	lbs, err := loadbalancers.ExtractLoadBalancers(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract LB list")
-	}
-	if len(lbs) > 1 {
-		return "", errors.Errorf("found multiple LB matching name %s, tag %s, cannot proceed", name, tag)
-	} else if len(lbs) == 1 {
-		if lbs[0].ProvisioningStatus == "ACTIVE" {
-			return lbs[0].ID, nil
-		} else if lbs[0].ProvisioningStatus == "ERROR" {
-			err := deleteOpenStackLb(client, lbs[0].ID)
-			if err != nil {
-				return "", errors.Wrap(err, "failed to delete LB")
-			}
-		}
-	}
-
-	createOpts := loadbalancers.CreateOpts{
-		Name:        name,
-		VipAddress:  vipAddress,
-		VipSubnetID: vipSubnetId,
-	}
-	if octaviaTagSupport {
-		createOpts.Tags = []string{tag}
-	} else {
-		createOpts.Description = tag
-	}
-	lb, err := loadbalancers.Create(client, createOpts).Extract()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to create LB")
-	}
-	err = waitForOpenStackLb(client, lb.ID)
-	if err != nil {
-		return "", errors.Errorf("Timed out waiting for the LB %s to become ready", lb.ID)
-	}
-
-	return lb.ID, nil
-}
-
-// Looks for a Octavia load balancer pool by name and LB ID. If it does
-// not exist creates it. Will fail if multiple LB pools are matching all criteria.
-func ensureOpenStackLbPool(client *gophercloud.ServiceClient, name, lbId string) (string, error) {
-	page, err := pools.List(client, pools.ListOpts{
-		Name:           name,
-		LoadbalancerID: lbId,
-		Protocol:       "HTTPS",
-		LBMethod:       "ROUND_ROBIN",
-	}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get LB pools list")
-	}
-	poolsList, err := pools.ExtractPools(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract LB pools list")
-	}
-	if len(poolsList) > 1 {
-		return "", errors.Errorf("found multiple LB pools matching name %s, LB %s, cannot proceed", name, lbId)
-	} else if len(poolsList) == 1 {
-		return poolsList[0].ID, nil
-	} else {
-		opts := pools.CreateOpts{
-			Name:           name,
-			LoadbalancerID: lbId,
-			Protocol:       pools.ProtocolHTTPS,
-			LBMethod:       pools.LBMethodRoundRobin,
-		}
-		poolsObj, err := pools.Create(client, opts).Extract()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to create LB pool")
-		}
-
-		err = waitForOpenStackLb(client, lbId)
-		if err != nil {
-			return "", errors.Errorf("Timed out waiting for the LB %s to become ready", lbId)
-		}
-
-		return poolsObj.ID, nil
-	}
-}
-
-// Looks for Octavia load balancer health monitor by name and pool ID. If it does
-// not exist creates it. Will fail if multiple LB health monitors are matching all criteria.
-func ensureOpenStackLbMonitor(client *gophercloud.ServiceClient, name, poolId string) (string, error) {
-	octaviaHTTPSMonitors, err := IsOctaviaVersionSupported(client, MinOctaviaVersionWithHTTPSMonitors)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to determine if Octavia supports HTTPS health monitors")
-	}
-
-	page, err := monitors.List(client, monitors.ListOpts{
-		Name:   name,
-		PoolID: poolId,
-	}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get LB monitors list")
-	}
-	monitorsList, err := monitors.ExtractMonitors(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract LB monitors list")
-	}
-	if len(monitorsList) > 1 {
-		return "", errors.Errorf("found multiple LB monitors matching name %s, pool %s, cannot proceed", name, poolId)
-	} else if len(monitorsList) == 1 {
-		return monitorsList[0].ID, nil
-	} else {
-		opts := monitors.CreateOpts{
-			Name:       name,
-			PoolID:     poolId,
-			Type:       monitors.TypeTCP,
-			MaxRetries: 3,
-			Delay:      10,
-			Timeout:    10,
-		}
-		if octaviaHTTPSMonitors {
-			// TODO(dulek): Octavia is a wild animal. So in OpenStack Stein the meaning of HTTPS type of health monitor
-			//              changed. Before, it was a simple TLS handshake. Now it's really doing an HTTPS connection to
-			//              a certain URL and TLS-HELLO is the new HTTPS. We're going to stick with the simple TCP check
-			//              for Octavia's prior to Stein and in Stein we switch to proper HTTPS healthcheck. The former
-			//              behavior can be removed once we stop supporting OpenStack Queens and Rocky.
-			opts.URLPath = "/healthz"
-			opts.Type = monitors.TypeHTTPS
-		}
-		monitorObj, err := monitors.Create(client, opts).Extract()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to create LB monitor")
-		}
-
-		return monitorObj.ID, nil
-	}
-}
-
-// Looks for a Octavia load balancer pool member by name, address and port. If
-// it does not exist creates it. Will fail if multiple LB pool members are
-// matching all criteria.
-func ensureOpenStackLbPoolMember(client *gophercloud.ServiceClient, name, lbId, poolId,
-	address, subnetId string, port, weight int) (string, error) {
-	page, err := pools.ListMembers(client, poolId, pools.ListMembersOpts{
-		Name:         name,
-		Address:      address,
-		ProtocolPort: port,
-	}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get LB member list")
-	}
-	members, err := pools.ExtractMembers(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract LB members list")
-	}
-	if len(members) > 1 {
-		return "", errors.Errorf("found multiple LB members matching name %s, cannot proceed", name)
-	} else if len(members) == 1 {
-		return members[0].ID, nil
-	} else {
-		opts := pools.CreateMemberOpts{
-			Name:         name,
-			Address:      address,
-			ProtocolPort: port,
-			SubnetID:     subnetId,
-			Weight:       &weight,
-		}
-		poolsObj, err := pools.CreateMember(client, poolId, opts).Extract()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to create LB member")
-		}
-
-		err = waitForOpenStackLb(client, lbId)
-		if err != nil {
-			return "", errors.Errorf("Timed out waiting for the LB %s to become ready", lbId)
-		}
-
-		return poolsObj.ID, nil
-	}
-}
-
-// Looks for a Octavia load balancer listeners by name, port, pool ID and LB ID.
-// If it does not exist creates it. Will fail if multiple LB listeners are
-// matching all criteria.
-func ensureOpenStackLbListener(client *gophercloud.ServiceClient, name, lbId, poolId string, port int) (string, error) {
-	page, err := listeners.List(client, listeners.ListOpts{
-		Name:           name,
-		Protocol:       "HTTPS",
-		ProtocolPort:   port,
-		DefaultPoolID:  poolId,
-		LoadbalancerID: lbId,
-	}).AllPages()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get LB listeners list")
-	}
-	listenersList, err := listeners.ExtractListeners(page)
-	if err != nil {
-		return "", errors.Wrap(err, "failed to extract LB listeners list")
-	}
-	if len(listenersList) > 1 {
-		return "", errors.Errorf("found multiple LB listeners matching name %s, LB %s, cannot proceed", name, lbId)
-	} else if len(listenersList) == 1 {
-		return listenersList[0].ID, nil
-	} else {
-		opts := listeners.CreateOpts{
-			Name:           name,
-			Protocol:       listeners.ProtocolHTTPS,
-			ProtocolPort:   port,
-			DefaultPoolID:  poolId,
-			LoadbalancerID: lbId,
-		}
-
-		// NOTE(dulek): If Octavia supports setting data timeouts in listeners (Rocky+) we set them to 10 mins hour as this
-		//              LB will be used for watching the Kubernetes API, that shouldn't time out after the default 50 seconds.
-		timeoutSupport, err := IsOctaviaVersionSupported(client, MinOctaviaVersionWithTimeouts)
-		if err != nil {
-			return "", errors.Wrap(err, "failed to determine if Octavia supports listener timeouts API")
-		}
-		timeout := 600000
-		if timeoutSupport {
-			opts.TimeoutClientData = &timeout
-			opts.TimeoutMemberData = &timeout
-		}
-
-		listenerObj, err := listeners.Create(client, opts).Extract()
-		if err != nil {
-			return "", errors.Wrap(err, "failed to create LB listener")
-		}
-
-		err = waitForOpenStackLb(client, lbId)
-		if err != nil {
-			return "", errors.Errorf("Timed out waiting for the LB %s to become ready", lbId)
-		}
-
-		return listenerObj.ID, nil
-	}
-}
-
-func getProjectID(keystone *gophercloud.ServiceClient, username, projectName string) (string, error) {
-	tokenID := keystone.Token()
-	proj, err := tokens.Get(keystone, tokenID).ExtractProject()
-	if err != nil {
-		return "", errors.Wrap(err, "failed to get token")
-	}
-	return proj.ID, nil
 }
 
 func generateName(name, clusterID string) string {
@@ -921,7 +248,7 @@ func BootstrapKuryr(conf *operv1.NetworkSpec, kubeClient client.Client) (*bootst
 			return nil, errors.Wrap(err, "failed to create Keystone client")
 		}
 
-		projectID, err := getProjectID(keystone, cloud.AuthInfo.Username, cloud.AuthInfo.ProjectName)
+		projectID, err := getProjectID(keystone)
 		if err != nil {
 			return nil, errors.Wrap(err, "failed to find project ID")
 		}
@@ -1174,15 +501,11 @@ func BootstrapKuryr(conf *operv1.NetworkSpec, kubeClient client.Client) (*bootst
 	} else {
 		octaviaProvider = "default"
 		if octaviaProviderSupport {
-			page, err := providers.List(lbClient, providers.ListOpts{}).AllPages()
+			providerList, err := listOpenStackOctaviaProviders(lbClient)
 			if err != nil {
 				log.Print("failed to get lbs provider list, using default octavia provider")
 			} else {
-				providers, err := providers.ExtractProviders(page)
-				if err != nil {
-					return nil, errors.Wrap(err, "failed to extract providers list")
-				}
-				for _, provider := range providers {
+				for _, provider := range providerList {
 					if provider.Name == OVNProvider {
 						log.Print("OVN Provider is enabled and Kuryr will use it")
 						octaviaProvider = OVNProvider
